refactor(sdk-go): split agent lifecycle methods
Move lifecycle and server orchestration methods out of agent.go into agent_lifecycle.go with no behavior changes.

Made-with: Cursor
